フォーマット処理を専用ファイルに分離
